Contains dynamic library source
This source contains code for "exaggerating" the colours of each pixel. For example, if the colours are (20, 40, 60) (assume B G R format), then the following operation is done on the pixel to increase the effect of the colour containing a greater value.

The updated pixel's expression:
(B^2 / (B^2 + G^2 + R^2) ) * 255,  (G^2 / (B^2 + G^2 + R^2)) * 255,  (R^2 / (B^2 + G^2 + R^2))*255 ).

Because squaring the pixels increases their magnitude disproportionally (with a greater value gaining an even greater value), The ratio of the value's square and the sum of squares of all pixel values, tend to be greater when the value's initial size was large.